Refactor: service layer
서비스 인프라에 관한 인터페이스 추가.

+++ src/services/user-service.ts
@@ -1,42 +1,47 @@
 import { cardClient, userClient } from '@/client';
 import { CardDto } from '@/domain/card';
 import { User } from '@/domain/user';
-import { UserDto } from '@/domain/user/user-dto';
+import { UserService } from './types';
 
-export const userService = {
-  get: async (id: string): Promise<UserDto | null> => {
+const isCardDto = (obj: any): obj is CardDto => {
+  return (
+    obj.id ||
+    obj.title ||
+    obj.phone ||
+    obj.message ||
+    obj.scan ||
+    obj.updatedAt ||
+    obj.updatedBy ||
+    obj.deviceId ||
+    obj.ownerList
+  );
+};
+
+export const userService: UserService = {
+  get: async id => {
     const user = await userClient.getById(id);
     return user;
   },
-  create: async (
-    user: { id: string } & Partial<
-      Pick<UserDto, 'cardIdList' | 'id' | 'nickname' | 'phone'>
-    >,
-  ): Promise<UserDto | null> => {
+  create: async user => {
     const dto = User.create(user).toDto();
     const res = await userClient.create(dto);
     if (!res) return null;
     return dto;
   },
-  updateNicknameAndPhone: async (
-    id: string,
-    nickname: string,
-    phone: string,
-  ): Promise<boolean> => {
+  updateNicknameAndPhone: async (id, nickname, phone) => {
     return await userClient.update({ id, nickname, phone });
   },
-  updateCardIdList: async (
-    id: string,
-    cardList: string[],
-  ): Promise<boolean> => {
+  updateCardIdList: async (id, cardList) => {
     return await userClient.update({ id, cardIdList: cardList });
   },
-  delete: async (id: string): Promise<boolean> => {
+  delete: async id => {
     const user = await userClient.getById(id);
     if (!user) return false;
 
     user.cardIdList.forEach(async cardId => {
-      const card = (await cardClient.getById(cardId)) as CardDto;
+      const card = await cardClient.getById(cardId);
+      if (!isCardDto(card)) return;
+      
       card.ownerList = card.ownerList.filter(userId => userId !== id);
       if (card.ownerList.length === 0) await cardClient.deleteById(cardId);
       else await cardClient.update(card);
@@ -44,7 +49,7 @@ export const userService = {
 
     return await userClient.deleteById(id);
   },
-  deleteCard: async (userId: string, cardId: string): Promise<boolean> => {
+  deleteCard: async (userId, cardId) => {
     const user = await userClient.getById(userId);
     const card = await cardClient.getById(cardId);
     if (!user || !card) return false;
